Adding error handling when not able to unpickle data
This is especially relavent given the previous commit, which involves
implementing unicode behavior for all responses.  Because pickling will try to
convert the pickled data to a string, it will fail when trying to interpret it
as unicode.

In this case, the only suggestion is to remove ``sessions.dat``, which is
mentioned in the exception.